Get rid of binprefixes.
They were not documented anyway.
And they were used in the wrong way, being taken
from the node connected to instead of from the
connecting node.  If diffent nodes have different
prefixes this fails horribly.

In the new setup, mclient must either be on the path
or $MCLIENT_PREFIX must be set to a path ending in a slash.